(feat) db connection, file reordering, quotation marks change
