+++ app.js
@@ -1,0 +1,48 @@
+/* ===== ENVIROMENT VARIBALES FILE CONFIG ===== */
+
+require('dotenv').config({ path: '.env' })
+
+/* ========== */
+
+/* ===== REQUIRED IMPORTS  ===== */
+
+const express = require('express')
+const mongoose = require('mongoose')
+const config = require('./config')
+
+/* ========== */
+
+/* ===== VARIABLES ===== */
+
+const PORT = process.env.PORT || 8082
+
+/* ========== */
+
+/* ===== APP INITIALIZATION ===== */
+
+const app = express()
+
+/* =========== */
+
+/* ===== DATABASE CONNECTION ===== */
+
+mongoose.connect(`${config.atlas.SCHEMA}://${config.atlas.USER}:${config.atlas.PASSWORD}@${config.atlas.HOSTNAME}/${config.atlas.DATABASE}?${config.atlas.OPTIONS}`).then(() => {
+
+    /* ===== MIDDLEWWARES ===== */
+
+    app.use(express.json())
+    app.use(express.urlencoded({ extended: true }))
+
+    /* ========== */
+
+    /* ===== APP LISTENING ===== */
+
+    app.listen(PORT, () => console.log(`Listening on port ${PORT}`))
+
+    /* ========== */
+
+})
+
+/* ========== */
+
+
