+++ config/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+    atlas: {
+        SCHEMA: process.env.ATLASSCHEMA,
+        USER: process.env.ATLASUSER,
+        PASSWORD: process.env.ATLASPASSWORD,
+        DATABASE: process.env.ATLASDATABASE,
+        OPTIONS: process.env.ATLASOPTIONS,
+        HOSTNAME: process.env.ATLASHOSTNAME
+    }
+}
