--- index.js
@@ -1,37 +1,0 @@
-function laCajaDePandora(numero) {
-    // proximamente escribiremos codigo aqui
-
-    if(numero % 2 === 0) return (numero >>> 0).toString(2)
-
-    return (numero >>> 0).toString(16)
-
-}
-
-function sergio() {
-    return { nombre: 'Carlitos Balá', edad:31, nacionalidad:'Argentina' }
-}
-
-
-function luciano() {
-    return {nombre: 'Larri Capija', edad: 21, nacionalidad: 'Argentina'}
-}
-
-function Angela() {
-    return { nombre: 'Elsa Pallo', edad:25, nacionalidad:'Narnia'
-    }
-}
-
-function martin(){
-    return {nombre: 'Martin', edad: 20, nacionalidad: 'Argentina'}
-}
-
-
-function Belen(){
-    return {nombre: 'Belen', edad:18, nacionalidad: 'Argentina'}
-}
-
-function andres(){
-    return {nombre: 'Anthony Lincoln', edad: 26, nacionalidad: 'Argentina'}
-}
-
-
